+++ CHANGELOG.md
@@ -1,26 +1,40 @@
+## 1.2.0 - Relative
+
+* Support relative imports (by default).
+* Use importmagic3 if installed.
+* Format code with prettier
+
 ## 1.1.1 - Contained
+
 * Restrict commands to python editors
 
 ## 1.1.0 - Oops
+
 * Fix wrong python file test
 
 ## 1.0.0 - ES6
+
 * ES6 rewrite
 * Use the project of the current file
 
 ## 0.4.0 - Verbose
+
 * Prevent inserting imports before top level comments
 
 ## 0.3.0 - Complex
+
 * Complete more complex prefixes (os.path)
 
 ## 0.2.1 - Indentation
+
 * Fix indentation
 
 ## 0.2.0 - Better release
+
 * Add an autocomplete provider
 * Use isort
 
 ## 0.1.0 - First Release
+
 * Every feature added
 * Every bug fixed